added resources and various fixes

+++ cmd/cleanup.php
@@ -14,7 +14,16 @@ echo( "Colourspace Connection Returner \n");
 
 try {
 
-    echo("- Getting files in temp folder");
+    echo("- Getting minutes");
+
+    if( count( $argv ) !== 1 )
+        $minutes = $argv[1];
+    else
+        $minutes = 5;
+
+    echo("\n --> " . $minutes );
+
+    echo("\n- Getting files in temp folder");
 
     $files = glob(COLOURSPACE_ROOT . UPLOADS_TEMPORARY_DIRECTORY . "*" );
 
@@ -25,6 +34,13 @@ try {
 
         if( is_dir( $file ) )
             continue;
+
+        //10 minutes
+        if(  filemtime ( $file ) > time() - ( 60 * $minutes ) )
+        {
+            echo("\n --> Skipping file: " . $file . " ( since too new )");
+            continue;
+        }
 
         echo( "\n --> Removing file: " . $file );
 
@@ -48,6 +64,12 @@ try {
         if( is_dir( $file ) )
             continue;
 
+        if(  filemtime ( $file ) > time() - ( 60 * $minutes ) )
+        {
+            echo("\n --> Skipping file: " . $file . " ( since too new )");
+            continue;
+        }
+
         echo( "\n --> Removing file: " . $file );
 
         if(@unlink( $file ) == false )
@@ -56,7 +78,7 @@ try {
         $count2++;
     }
 
-    echo("\n- Completed. Deleted total: " . $count );
+    echo("\n- Completed. Deleted total: " . $count2 );
 
     echo("\n- All files removed. Total removed: " . ( $count + $count2 ) );
 }
@@ -66,4 +88,4 @@ catch ( Error $error )
     echo( "[Critical Error] : " . $error->getMessage() . "\n" );
 }
 
-echo("Finished");
+echo("\nFinished");
